+++ gluten-data/src/main/java/org/apache/gluten/columnarbatch/IndicatorVectorBase.java
@@ -16,7 +16,7 @@
  */
 package org.apache.gluten.columnarbatch;
 
-import org.apache.gluten.exec.Runtimes;
+import org.apache.gluten.runtime.Runtimes;
 
 import org.apache.spark.sql.types.DataTypes;
 import org.apache.spark.sql.types.Decimal;
