+++ gluten-data/src/main/java/org/apache/gluten/datasource/DatasourceJniWrapper.java
@@ -16,8 +16,8 @@
  */
 package org.apache.gluten.datasource;
 
-import org.apache.gluten.exec.Runtime;
-import org.apache.gluten.exec.RuntimeAware;
+import org.apache.gluten.runtime.Runtime;
+import org.apache.gluten.runtime.RuntimeAware;
 import org.apache.gluten.init.JniUtils;
 
 import org.apache.spark.sql.execution.datasources.BlockStripes;
