+++ gluten-data/src/main/java/org/apache/gluten/runtime/RuntimeAware.java
@@ -14,22 +14,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.gluten.exec
+package org.apache.gluten.runtime;
 
-import org.apache.spark.util.{TaskResource, TaskResources}
-
-object Runtimes {
-
-  /** Get or create the runtime which bound with Spark TaskContext. */
-  def contextInstance(name: String): Runtime = {
-    if (!TaskResources.inSparkTask()) {
-      throw new IllegalStateException("This method must be called in a Spark task.")
-    }
-
-    TaskResources.addResourceIfNotRegistered(name, () => create(name))
+/**
+ * This defines the base abstraction for the contextual objects that can be transmitted to C++ side
+ * for further native processing.
+ */
+public interface RuntimeAware {
+  default boolean isCompatibleWith(RuntimeAware other) {
+    return handle() == other.handle();
   }
 
-  private def create(name: String): Runtime with TaskResource = {
-    Runtime(name)
-  }
+  long handle();
 }
