+++ gluten-data/src/main/java/org/apache/gluten/vectorized/NativePlanEvaluator.java
@@ -17,8 +17,8 @@
 package org.apache.gluten.vectorized;
 
 import org.apache.gluten.backendsapi.BackendsApiManager;
-import org.apache.gluten.exec.Runtime;
-import org.apache.gluten.exec.Runtimes;
+import org.apache.gluten.runtime.Runtime;
+import org.apache.gluten.runtime.Runtimes;
 import org.apache.gluten.memory.memtarget.MemoryTarget;
 import org.apache.gluten.memory.memtarget.Spiller;
 import org.apache.gluten.memory.memtarget.Spillers;
