+++ gluten-data/src/main/java/org/apache/gluten/vectorized/NativeRowToColumnarJniWrapper.java
@@ -16,8 +16,8 @@
  */
 package org.apache.gluten.vectorized;
 
-import org.apache.gluten.exec.Runtime;
-import org.apache.gluten.exec.RuntimeAware;
+import org.apache.gluten.runtime.Runtime;
+import org.apache.gluten.runtime.RuntimeAware;
 
 public class NativeRowToColumnarJniWrapper implements RuntimeAware {
   private final Runtime runtime;
