+++ gluten-data/src/main/java/org/apache/gluten/vectorized/PlanEvaluatorJniWrapper.java
@@ -16,8 +16,8 @@
  */
 package org.apache.gluten.vectorized;
 
-import org.apache.gluten.exec.Runtime;
-import org.apache.gluten.exec.RuntimeAware;
+import org.apache.gluten.runtime.Runtime;
+import org.apache.gluten.runtime.RuntimeAware;
 import org.apache.gluten.validate.NativePlanValidationInfo;
 
 /**
