+++ gluten-data/src/main/scala/org/apache/gluten/runtime/Runtimes.scala
@@ -14,16 +14,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.gluten.exec;
+package org.apache.gluten.runtime
 
-/**
- * This defines the base abstraction for the contextual objects that can be transmitted to C++ side
- * for further native processing.
- */
-public interface RuntimeAware {
-  default boolean isCompatibleWith(RuntimeAware other) {
-    return handle() == other.handle();
+import org.apache.spark.util.{TaskResource, TaskResources}
+
+object Runtimes {
+
+  /** Get or create the runtime which bound with Spark TaskContext. */
+  def contextInstance(name: String): Runtime = {
+    if (!TaskResources.inSparkTask()) {
+      throw new IllegalStateException("This method must be called in a Spark task.")
+    }
+
+    TaskResources.addResourceIfNotRegistered(name, () => create(name))
   }
 
-  long handle();
+  private def create(name: String): Runtime with TaskResource = {
+    Runtime(name)
+  }
 }
